+++ tests/s3_tester.h
@@ -121,7 +121,7 @@ struct aws_s3_tester {
 
 struct aws_s3_tester_client_options {
     enum aws_s3_client_tls_usage tls_usage;
-    uint64_t part_size;
+    size_t part_size;
     size_t max_part_size;
     uint32_t setup_region : 1;
     uint32_t use_proxy : 1;
